requirements.txt
av==14.0.1
contourpy==1.3.1
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
imageio==2.36.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.5
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
psutil==6.1.1
pygame==2.6.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
setuptools==75.6.0
six==1.17.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
triton==3.1.0
typing_extensions==4.12.2
